==> notebooks/00_data_import.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split


df = pd.read_csv(
    r'C:\Users\sacii\OneDrive\Desktop\Oslomet\ACIT4510 Statistical learning\ACIT4510_statistical_learning_project\data\raw\Mental_Health_Lifestyle_Dataset.csv'
)


pd.set_option('display.max_columns', None)  # Show all columns
pd.set_option('display.max_rows', None)     # Show all rows

df.shape


df.head(5)

##### Stratifying on binned Happiness Score keeps the target distribution consistent across train, eval, and test sets. Since the score is continuous, qcut bins it into quantiles so stratification works. This prevents imbalanced splits, reduces bias, and stabilizes model evaluation by ensuring all subsets represent the same range of happiness values.

strat = pd.qcut(df["Happiness Score"], q=10, duplicates="drop")

train_df, test_df = train_test_split(df, test_size=0.20, random_state=42, stratify=strat)

strat_train = pd.qcut(train_df["Happiness Score"], q=10, duplicates="drop")

train_df, eval_df = train_test_split(train_df, test_size=0.20, random_state=42, stratify=strat_train)

save_path = r'C:\Users\sacii\OneDrive\Desktop\Oslomet\ACIT4510 Statistical learning\ACIT4510_statistical_learning_project\data\raw'

train_df.to_csv(f"{save_path}\\train_df.csv", index=False)
test_df.to_csv(f"{save_path}\\test_df.csv", index=False)
eval_df.to_csv(f"{save_path}\\eval_df.csv", index=False)

print("Training set:", train_df.shape)
print("Evaluation set:", eval_df.shape)
print("Test set:", test_df.shape)
